Delete core/tables/queries.js, shove what's left of the legacy query junk in to TableClass.js
